requirements.txt
Flask==3.1.1
pandas==2.3.1
numpy==2.3.1
python-dotenv==1.1.1
requests==2.32.4
gunicorn==23.0.0
